Really basic date parsing for CLI
You can now call

    wm import ia http://epa.gov --from 2017-03-01 --to 2017-04-01

to import only versions from a given timeframe. Only supported for IA imports. Note time zones are not really respected. Need to get a better handle on how IA manages them, then figure out how to manage them on the parsing side (we are leaning on Pandas for the date parsing, but it doesn't seem to be super smart about it -- it may be better to use a dedicated library like iso8601, arrow, or the like).