Move change styles in HTML diff to end of <head>
This caused character encoding errors on some pages. When `<meta charset="utf-8">` is used, it *has* to be the first thing in the <head> and the <head> has to be the first thing in the <html>. The way we lazily inserted new CSS at the start of the document clobbered this. Additionally this means these ins/del styles now override anything defined elsewhere in the <head> or linked CSS files because they come last.