Remove repeat blank lines from text diff
Up to two repeat blank lines are allowed (which helps to get some differentiation between parts of the page), but anything more is collapsed. This also changes the output of side-by-side text to be two text strings rather than a list of text strings (since we need to collapse *across* strings).

Fixes #98.