Clean up newspaper experiment.
This work is from the NYC Data Rescue event. We have reached out
to the original author in the past, and he does not have time
to regularly contribute.

The important content of this code was:

    from newspaper import article
    a = Article('FAKE_URL').set_html(raw_html)
    a.parse()
    a.text

We have an issue #44 to note that experimenting with the
newspaper module could be useful for extracting important content
from a page.

https://github.com/drdoctr/doctr/issues/189