Clean up issues encountered doing an IA import
While doing a one-time import of a few pages for the analyst team, I encountered several issues. I'm not sure how we didn't catch these in the initial DB client work.

Note some of these are good fixes, but others are probably not right (they are more like quick hacks to get it working). This may need some refinement.