html_token: stop early with an error for non-html
This is a simple attempt to resolve the big issues of asking the `html_token` differ to diff non-HTML content, which it simply can't do. It has minimal architectural changes, so it does its best to sniff the content its given to see if it it's almost certainly *not* HTML and stops early in that case (note this is super lenient; it could get passed lots of types of invalid content and it would still try to diff it -- it works best against PDFs and images).

It's an open question whether this should be more strict (e.g. should it sniff for things that explicitly look *like* HTML, not just things that *don't?*) or if we should make architectural changes to benefit it, e.g. stoping with text decoding errors or passing headers along for the diff algorithm to examine.

Fixes #147