Use minimal HTML in diff when there is no content
When there is no content (or just whitespace) for a given version, BeautifulSoup winds up creating no structure in the document at all (in all other situations, with any content present, it will create a root `<html>` element even if one is not present). Later on, we break when it turns out there is no root element to work with. Instead, this detects situations where there is just whitespace for content and substitutes an extremly simple HTML document for diffing. This has the added bonus of a little bit of text explaining there there was no content so analysts aren't confused about whether something is still loading, silently broken, or if there really was just no content.

Fixes #157.