Clean up the remaining major issue in links diff
This "fixes" the last issue I've been able to identify with the overall algorithm in the links diff, and unfortunately makes it really complicated. This last part makes me seriously wonder whether the way I'm abusing SequenceMatcher here is really the right approach. That said, it does seem to work.

The last major hurdle here is the fact that the loose equality checking for Links, which was meant to make SequenceMatcher classify sets of *similar* links as equal, whereupon we could do a second pass with more complicated checking, could fail if the number of similar links in the before or after version was different. Take this example list of links:

   A                             B
   --------------------------- | ------------------------------
1. "Pony time!" ponytime.com/a | "Pony time!" ponytime.com/b
2. "Pony time!" ponytime.com/b | "Donkey time." not-ponies.com/

What we *want* is to wind up with:

1. Removed "Pony time!" ponytime.com/a
2. Equal   "Pony time!" ponytime.com/b  =  "Pony time!" ponytime.com/b
3. Added   "Donkey time." not-ponies.com/

But because items A1, A2, and B1 are all "equal", SequenceMatcher groups A1 + B1 together as unchanged (good), but then makes A2 a removal. i.e. we got:

1. Equal "Pony time!" ponytime.com/a  = "Pony time!" ponytime.com/b (which we then diff the components of)
2. Removed "Pony time!" ponytime.com/b
3. Added "Donkey time." not-ponies.com/

That's *ok,* but ideally the removal should be A1, while A2 + B1 are the equal group (because A2 + B2 are *exactly* equal). So! This change basically does a first pass after SequenceMatcher to look at all the added/removed items and, if they are roughly equal to the items in the equal section immediately preceeding them (or following), *move* them from the added/removed to the equal section. That then means the equal sections will be disbalanced (i.e. they have more items in the A than B side or vice-versa), so the handling of the pass where we do deeper comparisons on the equal group also gets more complex.

I suspect there might be a better overall solution in finding sequences of roughly equal Links and bundling them up in some othe token *before* sending them to SequenceMatcher. It sounds more manageable, at least. But I am tired, that would be a lot of work to prototype, and what I've already done here seems to work, so pushing this for now.