Remove invisible content from link diff text
This leverages the HTML diff's `undiffable_content_tags` to find and remove non-text elements (like `<script>` or `<style>`) from the DOM so they don't show up in the text of a link. It also slightly improves the text algorithm when there's *no* text to show: we use the `title` attribute if there is one, otherwise straight-up *say* there's no text.

Fixes #198.