Return to being totally lenient on text decoding
This removes the `?ignore_decoding_errors` query param and changes the behavior to never raise exceptions when there are decoding problems (instead, we just replace the error with the unicode "REPLACEMENT CHARACTER" character, like a browser generally would).

This winds up partially reverting #155 and removing one of our protections against incorrectly trying to diff large binary blobs, but the other content sniffing logic added in #155 is still there and should hopefully be a good enough solution to that problem. The reality is that there's lots of content on the web with actual encoding errors, and we still need to be able to diff it. To do that, the UI would have turned on `ignore_decoding_errors` all the time anyway, so it was probably a pointless switch to have.

Fixes #203.