Improve element lookups using sets instead of lists
We have many lists of tags like `block_level_tags` and we check if an
element is inside them using `name in block_level_tags`.

When we do that and we use a list, Python iterates ALL elements make an
equal number of comparisons. If we use `set()` instead, Python performs
just one lookup.

Since this code is used a lot to check all these element, I think that there
will be a considerable performance improvement using `sets()`.