Pass diff request header to remote URLs
Add param `pass_header=HEADER_KEY` to diffing server.
When set, it gets HTTP header with specified key and passes it as
an HTTP header to remote URLs `a` and `b`.

Useful for passing things like cookies to remote URLs if necessary
(e.g. password protected sites with cookie authentication).

Please note that to use this feature with cookies in a web application,
you need an addition parameter `credentials: "include"` in your request.
https://fetch.spec.whatwg.org/#cors-protocol-and-credentials