Add header Access-Control-Allow-Credentials: true
With the latest new feature to pass HTTP headers to target pages,
we use `pass_headers=cookie` and AJAX fetch option `{credentials: 'include'}`
in our application to invoke `web-monitoring-processing`.

Unfortunately, the browser raises an error and requires also
`Access-Control-Allow-Credentials: true` in the response to work
properly. I'm adding this with the current PR.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials