Raise error decoding text if it has many errors
At one point, we were fairly strict on text decoding and stopped if there were any errors. Later, we loosened that because lots of content out there on the web has a few stray encoding failures (see #232). However, that opened us up to attempting to diff large binary blobs as text, which has *really* bad results.

As a middle ground solution, this brings back decoding errors, but only if the content is made up of > 25% errors. That's extremely unlikely in any document meant to be text (really, even if it was, that text would be totally garbled by this point). It gives us a decent heuristic to say that we can't treat this as meaningfully diffable text and gives us an easy way to bail on receiving binary content (like PDFs, images, etc.).

Fixes #287.