Detect text encoding using cchardet (#411)
`_extract_encoding` now uses cChardet to detect the encoding of content if it can't find an explicitly specified encoding.

It also always returns a supported codec name, even if that name was just a default. (It previously returned `None` if the encoding could not be detected.) This drastically simplifies its usage and fixes a potential situation where we might try decoding the body twice.

This also introduces a slight behavior change: we now fall back to UTF-8 when a specified/detected encoding is not supported. (Previously, we fell back to UTF-8 if not encoding was specified/detected, and ASCII if the specified/detected encoding was not supported.) This change isn't likely to have a major impact on accuracy, but does make the code easier to understand.