Use html5-parser instead of lxml (#410)
Use html5-parser instead of lxml when parsing HTML for diffs. (It still builds BeautifulSoup documents at the end of the day, so the code impact is minimal.) We think this should parse slightly faster, but more importantly, will do a better job parsing various malformed bits of HTML we encounter.

Because this needs to link with the version of lxml we use, installation gets a bit more complicated (See https://html5-parser.readthedocs.io/en/latest/). At the end of the day, pip can still automate most of it, but we now have some special instructions in the pipfile. If we ever distribute the diff code as a package, we’ll have to make special note of this because `setup.py` doesn’t support proper installation.