Make a separate `build_docker` job in CircleCI (#422)
In CI, this splits the image building part of the `publish_docker` job off into a `build_docker` job, and then runs that job alongside our normal tests. That means all our PRs will try to build (but not publish) Docker images, and we’ll know if there’s a problem with changes to the Dockerfile or some other part of the build process.

In #410, I did a bad job testing the Docker build (I tested an earlier version, but forgot to test the final version that we merged, which failed). That led to a hotfix (#419), which is not fun. This should help prevent that kind of situation.