Use more descriptive errors for upstream failures (#429)
When a request to another server to get diff content fails, we used to simply return the upstream server's error, which makes it tough to determine which request was problematic (was it the `a` or `b` param?) and can be hard to distinguish from an error inside the diffing server itself.

Instead, we now use 502/504 status codes in these cases (these codes are both for *gateways,* implying that the error happened upstream) and add the URL that failed to the error message. The upstream status code (if any) and message are also included in the full error text. This should help make it much more clear which request is failing and where exactly the failure is coming from.