Fix parsing body tags containing the string `</body>` (#478)
It turns out some totally broken diffs were because of a poorly thought-through set of regexes used to clean up documents. We use those at a particularly dumb point in the process where we serialize the document and re-parse it a second time (!) because the tokenization routines we inherited from lxml were not designed to work with BeautifulSoup nodes.

While we should eventually fix the tokenization and double-parsing, this smaller patch does away with the regexes. Since we already have a parsed BeautifulSoup representation, we can do the cleanup more easily and reliably on that before serializing for the second parse.

Fixes #259.