Remove PageFreezer support (#479)
Remove all our legacy, deprecated code for working with data and diffs from PageFreezer. We haven’t used it in a long time and it’s likely much of this code is buggy or out of date. We no longer maintain our own PageFreezer install anymore either, so it’s unlikely we’ll ever revisit this code in the near-to-medium future.

Fixes #379.