Make sure to close responses where they may leak (#21)
This is a bit of a follow-on to #19 and #20.

These are all the remaining places I could find possible connection leaks from responses being held onto. A lot of these cases *should* be covered by the requests library (it takes a similar approach and caches response bodies before raising most exceptions), but I didn't feel like I could be certain that was always true, so I went with the most cautious approach: anywhere we aren't explicitly returning an open request, read/cache the body and close it.