Update & specify sphinx and plugin versions
We had issues where the official docs generated by ReadTheDocs didn't match docs you'd generate locally because we failed to specify version requirements for Sphinx (and in the absence of a specified version ReadTheDocs will use a very outdated version). We also had some CSS customizations to fix the theme in Sphinx v2+ that did not work in v1. All of these are now fixed by specifying actual version requirements for Sphinx and by updating to the latest of Sphinx and the Sphinx plugins (like the ReadTheDocs theme).

Fixes #35.