Add failing tests for new get_memento signature
These tests will validate new possible values for the `url` parameter (and the new `date_` and `mode` parameters) in `get_memento`. The basic idea here is that the current interface, where you must pass an actual *memento URL* rather than the URL of the page you want a memento for, is pretty unintuitive for almost every case except EDGI's current usage. In other words, instead of doing:

    client.get_memento('https://web.archive.org/web/20200101000000id_/http://www.epa.gov/')

It's a lot easier and more sensible to do:

    client.get_memento('http://www.epa.gov/', date_=date(2020, 1, 1))

But we'll still support the old behavior above (both for backwards compatibility and because it's useful). In addition, since out main use case boils down to:

    for cdx_record in client.search('<some url>'):
        client.get_memento(cdx_record.raw_url)

We may as well allow the CDX record itself, too:

    for cdx_record in client.search('<some url>'):
        client.get_memento(cdx_record)

This is the first step towards #15.