Add `follow_redirects` parameter to `get_memento` (#49)
This adds a `follow_redirects` parameter to the `WaybackClient.get_memento()` method, which controls whether the client follows *historical* redirects to get the resulting memento of what would have been displayed in a browser. It defaults to `True`, which is the current behavior.

Note this is different than following Wayback Machine redirects that aren’t historical (i.e. the redirect is not a memento). These are already controlled by `exact`, `exact_redirects`, and `target_window`.

Fixes #14.