Stub out the Memento type
Instances of the `Memento` class will replace the return value from `WaybackClient.get_memento()`, which currently returns `requests.Response` objects. See #2 for more on plans here.

I've attempted to keep this relatively simple for now (e.g. headers are case-sensitive, content is not streamable/iterable) in order to minimize what needs to be done here to the essentials. We can then expand on it in later commits or PRs. I also expect this implementation to evolve at least a little over the course of the initial implementation of the feature.