Drop `media` attribute from `Memento` class
After investigating this more deeply, I realized proper parsing is a bit complex -- not exceedingly so, but enough that even `requests` takes shortcuts and will get it wrong in some edge-cases. (I think this might be why requests only uses it to determine the character encoding of the body and does not expose the full parsed media type.)

For now, I think it's better to not include this. EDGI needs it, but is already doing it a level up. That's less nice, but better to do questionable stuff there than here, where it's more widely used and the cost of incorrect behavior or a poor design is higher. I figure we'll come back around to it later.