Add `NoMementoError` for URLs with no mementos
The Wayback Machine returns a 404 error on Memento requests only when a URL was never actually archived (and maybe when no mementos are available for playback, but I think I've usually seen that become an infinite redirect loop instead). Raise a specific error for this scenario.

See #64 for more about errors we are trying to handle better.