Ensure session timeout is used, add default (#68)
This fixes a major bug where a WaybackSession's `timeout` would never actually be applied. It also adds a default value for the timeout (60 seconds for connections and reads). This is intentionally high, since the goal is only to provide guardrails to help keep your program for hanging endlessly (occasionally, the Wayback Machine just does not hang up on a connection).

Fixes #66.