Correct CDX API timestamps with month of 00
This commit extends the existing logic for handling invalid days of `00` to months that are `00`. It also adds a warning to be logged in both situations.

So if a timestamp of `20200001120000` will get rewritten to `20200112000000` prior to conversion to a datetime.

I have tested on live CDX API data that was failing, and this fix causes the
full result to be returned. If more information is known about why this
approach is taken it would be good to add in a comment?

Closes #88