Repair docs build (#91)
This updates a number of dev dependencies related to docs and pins them appropriately so we don't have future problems like the one that prompted this (an update to Jinja2 broke). There are still some unpinned dev dependencies that I should add constraints to in a follow-up change.

The biggest deal here is that the docs dependencies are now listed in a separate file, since they may conflict with other dev dependencies on certain versions of Python. (This is critical, since we need to run tests and verify support on Python 3.6+, and the docs dependencies can’t be installed alongside the other dev dependencies in Python < 3.8.)

This also changes how we do setup for Read the Docs. The old way has actually been deprecated (even though they log no warnings or other notices) and doesn’t work for Python versions newer than 3.8!