Minor performance improvement to timestamp fix (#92)
I had wondered before about the performance implications of making an extra list when parsing timestamps, and @edsu’s recent patch in #89 got me to take a look.

This is a minor 5-6% performance improvement, which may be worthwhile if you are iterating through extremely large result sets. Part of this is not allocating a few new lists, and part is checking the least significant digit first, since there are relatively few valid months/days where the least significant digit is 0.