Use snake_case for all `search()` parameters
We originally used camelCase when the name and functionality were the same as a parameter in the HTTP API. It was intended to make things clear and to support arbitrarily specifying additional keyword arguments that would get sent in the request. However, we no longer support additional, arbitrary keyword arguments, and the mixed case created more inconvenience and mess than help. This makes all parameters use the same snake_case scheme.