Defer assertions to end of timing tests
Otherwise it's a huge pain to get a valid cassette, since each test run only gets part of the cassette built (the time taken for the network request causes the test to fail if not read from a cassette, and if assertions are in the middle, requests after the failed assertions never get made and recorded, then the next run that lets those assertions pass can't overwrite the previously written cassette...).

This also loosens the expected timings in the test, since they just don't always get hit exactly owing to whatever else might be happening on the machine when tests run.