Deprecate instead of remove renamed parameters
In #94 and #102, we renamed several method parameters. This brings back the old names, but with deprecation warnings. We'll remove them fully in v0.5.0. (This does *not* support parameters that were completely *removed*, though. I did a search on GitHub, and it doesn't look like anybody is actually using them, so it didn't seem worth keeping support for them.)