Handle view-mode redirects
When requesting mementos of redirects in `view` mode, we get back a web page that redirects in JavaScript and that is missing some important memento headers, causing us to raise a pretty unexpected error. This attempts to work around the issue and detect that a page is a memento of a redirect.

This is a first pass, and probably needs some more cleanup.

Fixes #109.