Ensure `Memento.links` matches Memento's mode
The `links` attribute on a Memento object links to other related resources, such as the first/previous/next/last memento of the URL. These are super useful for iterating and navigating through Mementos, BUT the archive.org servers always return these as links to the memento in "view" mode. Users of this library rarely use "view" mode, and if you are iterating through links in a different mode, simply following one of the links can lead to subtle mistakes! To mitigate the issue, this changes all the links that reference mementos to use the same mode as the Memento object they are attached to.

Fixes #111.