Support Python 3.12 Prereleases (#125)
This adds CI test support for Python 3.12.0 pre-releases. It also adds setuptools as a test requirement. Setuptools is needed to do builds, but shouldn't be necessary as a general dependency because pip and other package managers maintain their own built-in copies for use during installation. (Long term, we should probably switch to TOML configuration and drop `setup.py`.)