Always delay at least 60 seconds for 429 responses
The 429 status code is a bit special -- the server is warning us we're going to get blocked if we don't slow it down. At current, the Wayback Machine has a 1 minute window for evaluating good behavior, so instead of pausing for the default backoff time, pause for at least 1 minute (if the default backoff time winds up being more, then pause for that).